add button-groups for `ProjectPropertiesTreeView`
This commit introduces a `ButtonGroupWidget` and it's parent `ButtonGroupDelegate` to the `project_properties` node. The button-group is intended to only be visible in the second column of the associated QTreeView. To accomodate this, the `JesterTreeViewModel` has been updated to accept a variable number of columns in it's constructor-method. This not as ideal as I would like but due to the nature of Qt's model/view system the model must contain the maximum amount of columns required by any views using it.

So far only the `add` and `remove` buttons have been implemented. During implementation of these methods I ran in to issues with the recursive methods on `JesterDirectory` objects and had to simplify the `path` and `project` methods.