feat: core logic update, FileCopyNode logic
This is a major update to the core API introducing a more robust data interface in preparation for usage in all nodes.

The initial `FileCopyNode` logic has been outlined which includes new logic for signalling upstream connections and disconnections of nodes. The intention is to include 2 new custom slots `on_upstream_node_connected` and `on_upstream_node_disconnected` which can be used to communicate changes in the connections upstream from any given node. The `FileCopyNode` uses these slots to keep a populated list of all `MediaSourceNode`'s.